chore: Move ProfileImage and SocialLinks components to `src/components/ui`
